requirements.txt
ctransformers
sentence-transformers
chromadb
pinecone-client
langchain
flask
pypdf
python-dotenv

